ub,db file move to cps/models/
