+++ cps/config_sql.py
@@ -28,11 +28,7 @@ from sqlalchemy import exists
 from cryptography.fernet import Fernet
 import cryptography.exceptions
 from base64 import urlsafe_b64decode
-try:
-    # Compatibility with sqlalchemy 2.0
-    from sqlalchemy.orm import declarative_base
-except ImportError:
-    from sqlalchemy.ext.declarative import declarative_base
+
 
 from . import constants, logger , utils
 from .subproc_wrapper import process_wait
@@ -44,158 +40,10 @@ import urllib.parse
 load_dotenv(utils.get_env_path())
 
 log = logger.create()
-_Base = declarative_base()
+# Settings models live in cps/models/settings.py (single source of truth).
+# Both this module and cps/forum/config.py import from there.
+from .models.settings import _Base, _Flask_Settings, _Settings  # noqa: F401
 
-
-class _Flask_Settings(_Base):
-    __tablename__ = 'flask_settings'
-
-    id = Column(Integer, primary_key=True)
-    flask_session_key = Column(BYTEA, default=b"")
-
-    def __init__(self, key):
-        super().__init__()
-        self.flask_session_key = key
-
-
-# Baseclass for representing settings in app.db with email server settings and Calibre database settings
-# (application settings)
-class _Settings(_Base):
-    __tablename__ = 'settings'
-
-    id = Column(Integer, primary_key=True)
-    mail_server = Column(String, default=constants.DEFAULT_MAIL_SERVER)
-    mail_port = Column(Integer, default=25)
-    mail_use_ssl = Column(SmallInteger, default=0)
-    mail_login = Column(String, default='mail@example.com')
-    mail_password_e = Column(String)
-    mail_password = Column(String)
-    mail_from = Column(String, default='automailer <mail@example.com>')
-    mail_size = Column(Integer, default=25*1024*1024)
-    mail_server_type = Column(SmallInteger, default=0)
-    mail_gmail_token = Column(JSON, default={})
-
-    config_calibre_dir = Column(String,default=os.environ.get('BOOK_FILEPATH'))
-    config_calibre_uuid = Column(String)
-    config_calibre_split = Column(Boolean, default=False)
-    config_calibre_split_dir = Column(String)
-    config_port = Column(Integer, default=constants.DEFAULT_PORT)
-    config_external_port = Column(Integer, default=constants.DEFAULT_PORT)
-    config_certfile = Column(String)
-    config_keyfile = Column(String)
-    config_trustedhosts = Column(String, default='')
-    config_calibre_web_title = Column(String, default='Calibre-Web')
-    config_books_per_page = Column(Integer, default=60)
-    config_random_books = Column(Integer, default=4)
-    config_authors_max = Column(Integer, default=0)
-    config_read_column = Column(Integer, default=0)
-    config_title_regex = Column(String,
-                                default=r'^(A|The|An|Der|Die|Das|Den|Ein|Eine'
-                                        r'|Einen|Dem|Des|Einem|Eines|Le|La|Les|L\'|Un|Une)\s+')
-    config_theme = Column(Integer, default=0)
-
-    config_log_level = Column(SmallInteger, default=logger.DEFAULT_LOG_LEVEL)
-    config_logfile = Column(String, default=logger.DEFAULT_LOG_FILE)
-    config_access_log = Column(SmallInteger, default=0)
-    config_access_logfile = Column(String, default=logger.DEFAULT_ACCESS_LOG)
-
-    config_uploading = Column(SmallInteger, default=1)
-    config_anonbrowse = Column(SmallInteger, default=1)
-    config_public_reg = Column(SmallInteger, default=1)
-    config_remote_login = Column(Boolean, default=False)
-    config_kobo_sync = Column(Boolean, default=False)
-
-    config_default_role = Column(SmallInteger, default=0)
-    config_default_show = Column(Integer, default=constants.ADMIN_USER_SIDEBAR)
-    config_default_language = Column(String(3), default="all")
-    config_default_locale = Column(String(2), default="en")
-    config_columns_to_ignore = Column(String)
-
-    config_denied_tags = Column(String, default="")
-    config_allowed_tags = Column(String, default="")
-    config_restricted_column = Column(SmallInteger, default=0)
-    config_denied_column_value = Column(String, default="")
-    config_allowed_column_value = Column(String, default="")
-
-    config_use_google_drive = Column(Boolean, default=False)
-    config_google_drive_folder = Column(String)
-    config_google_drive_watch_changes_response = Column(JSON, default={})
-
-    config_use_goodreads = Column(Boolean, default=False)
-    config_goodreads_api_key = Column(String)
-    config_register_email = Column(Boolean, default=True)
-    config_login_type = Column(Integer, default=2)
-
-    config_kobo_proxy = Column(Boolean, default=False)
-
-    config_ldap_provider_url = Column(String, default='example.org')
-    config_ldap_port = Column(SmallInteger, default=389)
-    config_ldap_authentication = Column(SmallInteger, default=constants.LDAP_AUTH_SIMPLE)
-    config_ldap_serv_username = Column(String, default='cn=admin,dc=example,dc=org')
-    config_ldap_serv_password_e = Column(String)
-    config_ldap_serv_password = Column(String)
-    config_ldap_encryption = Column(SmallInteger, default=0)
-    config_ldap_cacert_path = Column(String, default="")
-    config_ldap_cert_path = Column(String, default="")
-    config_ldap_key_path = Column(String, default="")
-    config_ldap_dn = Column(String, default='dc=example,dc=org')
-    config_ldap_user_object = Column(String, default='uid=%s')
-    config_ldap_member_user_object = Column(String, default='')
-    config_ldap_openldap = Column(Boolean, default=True)
-    config_ldap_group_object_filter = Column(String, default='(&(objectclass=posixGroup)(cn=%s))')
-    config_ldap_group_members_field = Column(String, default='memberUid')
-    config_ldap_group_name = Column(String, default='calibreweb')
-
-    config_kepubifypath = Column(String, default=None)
-    config_converterpath = Column(String, default=None)
-    config_binariesdir = Column(String, default=None)
-    config_calibre = Column(String)
-    config_rarfile_location = Column(String, default=None)
-    config_upload_formats = Column(String, default=','.join(constants.EXTENSIONS_UPLOAD))
-    config_unicode_filename = Column(Boolean, default=False)
-    config_embed_metadata = Column(Boolean, default=True)
-
-    config_updatechannel = Column(Integer, default=constants.UPDATE_STABLE)
-
-    config_reverse_proxy_login_header_name = Column(String)
-    config_allow_reverse_proxy_header_login = Column(Boolean, default=False)
-
-    schedule_start_time = Column(Integer, default=4)
-    schedule_duration = Column(Integer, default=10)
-    schedule_generate_book_covers = Column(Boolean, default=False)
-    schedule_generate_series_covers = Column(Boolean, default=False)
-    schedule_reconnect = Column(Boolean, default=False)
-    schedule_metadata_backup = Column(Boolean, default=False)
-
-    config_password_policy = Column(Boolean, default=True)
-    config_password_min_length = Column(Integer, default=8)
-    config_password_number = Column(Boolean, default=True)
-    config_password_lower = Column(Boolean, default=True)
-    config_password_upper = Column(Boolean, default=True)
-    config_password_character = Column(Boolean, default=True)
-    config_password_special = Column(Boolean, default=True)
-    config_session = Column(Integer, default=1)
-    config_ratelimiter = Column(Boolean, default=True)
-    config_limiter_uri = Column(String, default="")
-    config_limiter_options = Column(String, default="")
-    config_check_extensions = Column(Boolean, default=True)
-
-    # PostgreSQL Configuration Fields - ADD THESE
-    config_db_host = Column(String, default='localhost')
-    config_db_port = Column(String, default='5432')
-    config_db_name = Column(String, default='getmyebook_app')
-    config_db_user = Column(String, default='vasanth')
-    config_db_password_e = Column(String)
-    config_db_password = Column(String)
-    config_database_url = Column(String)
-    config_use_postgresql = Column(Boolean, default=False)
-    
-    # PostgreSQL Metadata Configuration
-    config_use_postgresql_metadata = Column(Boolean, default=False)
-    config_postgresql_metadata_url = Column(String)
-
-    def __repr__(self):
-        return self.__class__.__name__
 
 
 # Class holds all application specific settings in calibre-web
